+++ src/main/java/org/boutry/devops/HelloService.java
@@ -6,6 +6,6 @@ import javax.enterprise.context.ApplicationScoped;
 @ApplicationScoped
 public class HelloService {
     public String greeting(String name) {
-        return "hello " + name;
+        return "<h1>hello " + name + "</h1>";
     }
 }