+++ src/main/java/org/boutry/devops/config/DatabaseConfiguration.java
@@ -1,0 +1,8 @@
+package org.boutry.devops.config;
+
+import io.quarkus.arc.config.ConfigProperties;
+
+@ConfigProperties(prefix = "database")
+public interface DatabaseConfiguration {
+    String path();
+}