+++ src/test/java/org/boutry/devops/HelloResourceTest.java
@@ -18,4 +18,12 @@ public class HelloResourceTest {
              .body(is("hello"));
     }
 
+    @Test
+    public void testHelloName() {
+        given()
+                .when().get("/hello/Dumbo")
+                .then()
+                .statusCode(200)
+                .body(is("<h1>hello Dumbo</h1>"));
+    }
 }