+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Repo init
 - Quarkus web server generated
 - Hello + Hello (name) route create
+- Configuration as YAML
+- Docker configuration file mounted in /deployments/config/